# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokumentasi Mini Project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server Utama"
        }
    ],
    "paths": {
...
        "/api/v1/transaksi/penjualan": {
            "get": {
                "tags": [
                    "Transaksi Penjualan"
                ],
                "summary": "Get list of penjualan with pagination and search",
                "operationId": "c0b7f8c2d006c01f336cbbe5db91b57d",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by ID Nota or Kode Pelanggan",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Transaksi Penjualan"
                ],
                "summary": "Create a new penjualan transaction",
                "operationId": "d8c7c1ca56b05f796402db3410d311c1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tgl",
                                    "kode_pelanggan",
                                    "items"
                                ],
                                "properties": {
                                    "tgl": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-10"
                                    },
                                    "kode_pelanggan": {
                                        "type": "string",
                                        "example": "PLG-001"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "kode_barang": {
                                                    "type": "string",
                                                    "example": "B001"
                                                },
                                                "qty": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil menambahkan data penjualan"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/v1/transaksi/penjualan/{id}": {
            "get": {
                "tags": [
                    "Transaksi Penjualan"
                ],
                "summary": "Get detail of a penjualan transaction",
                "operationId": "a17244708d7df834fce7a0e5a27eda9f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of penjualan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Penjualan not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Transaksi Penjualan"
                ],
                "summary": "Update an existing penjualan transaction",
                "operationId": "c900965f21faad31729454d5f2319b0f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of penjualan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tgl": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-11"
                                    },
                                    "kode_pelanggan": {
                                        "type": "string",
                                        "example": "PLG-001"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "kode_barang": {
                                                    "type": "string",
                                                    "example": "B001"
                                                },
                                                "qty": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil mengubah data penjualan"
                    },
                    "404": {
                        "description": "Penjualan not found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transaksi Penjualan"
                ],
                "summary": "Delete a penjualan transaction",
                "operationId": "3966fbb5edd4a877100aa74a3e40e252",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of penjualan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil menghapus data penjualan"
                    },
                    "404": {
                        "description": "Penjualan not found"
                    }
                }
            }
        }
    },
...
}